+++ R/latex.R
@@ -46,14 +46,20 @@ to_latex <- function (ht, ...) UseMethod("to_latex")
 to_latex.huxtable <- function (ht, tabular_only = FALSE, ...){
   assert_that(is.flag(tabular_only))
   tabular <- build_tabular(ht)
-  if (tabular_only) return(maybe_markdown_fence(tabular))
+  commands <- "
+  \\providecommand{\\huxb}[2]{\\arrayrulecolor[RGB]{#1}\\global\\arrayrulewidth=#2pt}
+  \\providecommand{\\huxvb}[2]{\\color[RGB]{#1}\\vrule width #2pt}
+  \\providecommand{\\huxtpad}[1]{\\rule{0pt}{\\baselineskip+#1}}
+  \\providecommand{\\huxbpad}[1]{\\rule[-#1]{0pt}{#1}}\n"
+
+  if (tabular_only) return(maybe_markdown_fence(paste0(commands, tabular)))
 
   resize_box <- if (is.na(height <- height(ht))) c("", "") else {
     if (is.numeric(height)) height <- sprintf("%.3g\\textheight", height)
     c(sprintf("\\resizebox*{!}{%s}{", height), "}")
   }
 
-  cap <- if (is.na(cap <- caption(ht))) "" else {
+  cap <- if (is.na(cap <- make_caption(ht, "latex"))) "" else {
     hpos <- get_caption_hpos(ht)
     cap_setup <- switch(hpos,
       left   = "raggedright",
@@ -74,7 +80,8 @@ to_latex.huxtable <- function (ht, tabular_only = FALSE, ...){
   res <- paste0(resize_box[1], tabular, resize_box[2])
   res <- if (grepl("top", caption_pos(ht))) paste0(cap, lab, res) else paste0(res, cap, lab)
   res <- paste0(
-          sprintf("\\begin{table}[%s]\n",latex_float(ht)),
+          commands,
+          sprintf("\n\\begin{table}[%s]\n",latex_float(ht)),
           pos_text[1],
           "\n\\begin{threeparttable}\n",
           res,
@@ -143,9 +150,7 @@ build_tabular <- function(ht) {
   bg_colors <- format_color(bg_colors, default = "white")
   hhline_colors <- bg_colors
   hhline_colors[has_own_border] <- hb_colors[has_own_border]
-  hhline_colors_tex <- sprintf("[%s]", hhline_colors)
-  hhline_colors_tex[has_own_border & hb_default] <- ""
-  hhlines_horiz <- sprintf(">{\\huxb%s{%.4g}}%s", hhline_colors_tex, horiz_b, hb_chars)
+  hhlines_horiz <- sprintf(">{\\huxb{%s}{%.4g}}%s", hhline_colors, horiz_b, hb_chars)
   dim(hhlines_horiz) <- dim(horiz_b)
   no_hborder_in_row <- hb_maxes[row(hhlines_horiz)] == 0
   hhlines_horiz[no_hborder_in_row] <- ""
@@ -168,10 +173,8 @@ build_tabular <- function(ht) {
   vert_bc[no_lrb_b] <- rbind(NA, cbc$vert)[no_lrb_b] # vert border color above
   # if it's still NA, no border on any side had a defined colour
   # At the moment we reset to black. Otherwise maybe we "bleed" along the row from much earlier.
-  vert_bc_na <- is.na(vert_bc)
+
   vert_bc <- format_color(vert_bc, default = "black")
-  vert_bc_tex <- rep("", length(vert_bc))
-  vert_bc_tex[! vert_bc_na] <- sprintf("[%s]", vert_bc[! vert_bc_na])
   hhlines_vert <- rep("", length(vert_b))
   has_vert_b <- vert_b > 0
 
@@ -186,8 +189,8 @@ build_tabular <- function(ht) {
   vert_bchars[! vert_bs == "double" & ! has_horiz_b]  <- "|"
   vert_bchars[vert_bs == "double" & ! has_horiz_b]    <- "||"
 
-  hhlines_vert[has_vert_b] <- sprintf(">{\\huxb%s{%.4g}}%s",
-        vert_bc_tex[has_vert_b],
+  hhlines_vert[has_vert_b] <- sprintf(">{\\huxb{%s}{%.4g}}%s",
+        vert_bc[has_vert_b],
         vert_b[has_vert_b],
         vert_bchars[has_vert_b])
   hhlines_vert[vert_bchars == ""] <- ""
@@ -341,10 +344,9 @@ build_tabular <- function(ht) {
   bs_double <- cbs$vert == "double"
   bcol <- format_color(bcol, default = "black")
   bord_tex <- rep("", length(bord))
-  bcol_tex <- rep("", length(bcol))
-  bcol_tex[has_bcol] <- sprintf("[%s]", bcol[has_bcol])
+
   # bord_tex[has_bord] <- sprintf("!{%s\\vrule width %.4gpt}", bcol_tex[has_bord], bord[has_bord])
-  bord_tex[has_bord] <- sprintf("!{\\huxvb%s{%.4g}}", bcol_tex[has_bord], bord[has_bord])
+  bord_tex[has_bord] <- sprintf("!{\\huxvb{%s}{%.4g}}", bcol[has_bord], bord[has_bord])
   bord_tex[bs_double] <- paste0(bord_tex[bs_double], bord_tex[bs_double])
   dim(bord_tex) <- dim(cb$vert)
   # the first column is the left border of the left-most cell.
@@ -409,14 +411,7 @@ build_tabular <- function(ht) {
   colspec_top <- paste0(colspec_top, collapse = " ")
   colspec_top <- sprintf("{%s}\n", colspec_top)
 
-  commands <- "
-    \\providecommand{\\huxb}[2][0,0,0]{\\arrayrulecolor[RGB]{#1}\\global\\arrayrulewidth=#2pt}
-    \\providecommand{\\huxvb}[2][0,0,0]{\\color[RGB]{#1}\\vrule width #2pt}
-    \\providecommand{\\huxtpad}[1]{\\rule{0pt}{\\baselineskip+#1}}
-    \\providecommand{\\huxbpad}[1]{\\rule[-#1]{0pt}{#1}}
-  "
-  # use like \huxb[color]{width} in >{} sections
-  res <- paste0(commands, tenv_tex[1], width_spec, colspec_top, table_body, tenv_tex[2])
+  res <- paste0(tenv_tex[1], width_spec, colspec_top, table_body, tenv_tex[2])
   return(res)
 }
 
@@ -457,8 +452,7 @@ maybe_markdown_fence <- function (text) {
     in_rmarkdown <- ! is.null(knitr::opts_knit$get("rmarkdown.pandoc.to"))
     if (in_rmarkdown && requireNamespace("rmarkdown", quietly = TRUE)) {
       of <- rmarkdown::default_output_format(knitr::current_input())
-      fence <- rmarkdown::pandoc_version() >= "2.0.0" &&
-            "+raw_attribute" %in% of$options$md_extensions
+      fence <- rmarkdown::pandoc_version() >= "2.0.0"
     }
   }