+++ tests/testthat/test-set-interface.R
@@ -2,41 +2,16 @@
 context("Row and column functions, set_* interface")
 
 
-test_that("Row/column function examples unchanged", {
-  test_ex_same("every")
-  test_ex_same("final")
-})
-
-test_that("final", {
-  dfr <- data.frame(a = 1:5, b = 1:5, d = 1:5, e = 1:5)
-  expect_equivalent(final(2)(dfr, 1), 4:5)
-  expect_equivalent(final(2)(dfr, 2), 3:4)
-  expect_equivalent(final(6)(dfr, 1), 1:5)
-})
-
-
-test_that("every(), evens and odds", {
-  dfr <- data.frame(a = 1:7, b = 1:7, d = 1:7, e = 1:7)
-  expect_equivalent(every(3, from = 1)(dfr, 1), c(1, 4, 7))
-  expect_equivalent(every(3, from = 1)(dfr, 2), c(1, 4))
-  expect_equivalent(evens(dfr, 1), c(2, 4, 6))
-  expect_equivalent(evens(dfr, 2), c(2, 4))
-  expect_equivalent(odds(dfr, 1), c(1, 3, 5, 7))
-  expect_equivalent(odds(dfr, 2), c(1, 3))
-  expect_equivalent(everywhere(dfr, 1), 1:7)
-  expect_equivalent(everywhere(dfr, 2), 1:4)
-})
-
-
-test_that("set_cell_properties", {
-  ht <- huxtable(a = 1:5, b = letters[1:5], d = 1:5)
-  ht2 <- set_cell_properties(ht, 1, 1, font = "times", align = "right")
+test_that("Can use set_cell_properties", {
+  htxx <- huxtable(a = 1:5, b = letters[1:5], d = 1:5)
+  ht2 <- set_cell_properties(htxx, 1, 1, font = "times", font_size = 24)
   expect_equivalent(font(ht2)[1, 1], "times")
-  expect_equivalent(align(ht2)[1, 1], "right")
+  expect_equivalent(font_size(ht2)[1, 1], 24)
   r <- 2
   c <- 1
-  ht3 <- set_cell_properties(ht, r, c, bold = TRUE)
+  ht3 <- set_cell_properties(htxx, r, c, bold = TRUE, font = "times")
   expect_equivalent(bold(ht3)[2, 1], TRUE)
+  expect_equivalent(font(ht3)[2, 1], "times")
 
 })
 
@@ -159,34 +134,6 @@ test_that("set_all_*", {
 })
 
 
-test_that("map_all_*", {
-  ht <- huxtable(1:5, 5:1)
-
-  ht2 <- map_all_borders(ht, by_ranges(3, c(0, 1)))
-  expect_equivalent(left_border(ht2), 1 * (as.matrix(ht2) >= 3))
-  expect_equivalent(right_border(ht2), 1 * (as.matrix(ht2) >= 3))
-
-  ht3 <- map_all_border_colors(ht, by_ranges(3, c("red", "black")))
-  expected <- matrix(ifelse(as.matrix(ht) >= 3, "black", "red"), 5, 2)
-  expect_equivalent(left_border_color(ht3), expected)
-  expect_equivalent(right_border_color(ht3), expected)
-
-  ht4 <- map_all_border_styles(ht, by_ranges(3, c("solid", "double")))
-  expected <- matrix(ifelse(as.matrix(ht) >= 3, "double", "solid"), 5, 2)
-  expect_equivalent(left_border_style(ht4), expected)
-  expect_equivalent(right_border_style(ht4), expected)
-
-  ht5 <- map_all_padding(ht, by_ranges(3, c(0, 10)))
-  expect_equivalent(left_padding(ht5), 10 * (as.matrix(ht) >= 3))
-  expect_equivalent(right_padding(ht5), 10 * (as.matrix(ht) >= 3))
-
-  ht6 <- map_all_borders(ht, 1:2, 1:2, by_ranges(3, c(2, 4)))
-  expected <- matrix(c(2, 2, 0, 0, 0, 4, 4, 0, 0, 0), 5, 2)
-  expect_equivalent(left_border(ht6), expected)
-  expect_equivalent(right_border(ht6), expected)
-})
-
-
 test_that("set_all_* functions work when huxtable is not attached", {
   # NB as written this test can only be run from the command line; detach call silently fails
   library(huxtable)
@@ -246,6 +193,29 @@ test_that("set_outer_borders() works with non-standard/empty position arguments"
 })
 
 
+test_that("set_contents works", {
+  ht <- hux(1:3, 1:3)
+
+  expect_equivalent(set_contents(ht, 1:6), hux(1:3, 4:6))
+  expect_equivalent(set_contents(ht, 1:6, byrow = TRUE), hux(c(1, 3, 5), c(2, 4, 6)))
+  expect_equivalent(set_contents(ht, 1, 1, 0), hux(c(0, 2:3), 1:3))
+  expect_equivalent(set_contents(ht, 1, 1, 0), hux(c(0, 2:3), 1:3))
+  expect_equivalent(set_contents(ht, 2:3, 2, 3:2), hux(1:3, c(1, 3, 2)))
+
+  ht <- hux(a = 1:3, b = 1:3)
+  expect_equivalent(set_contents(ht, 1, "a", 0), set_contents(ht, 1, 1, 0))
+  # dplyr::matches not testthat::matches
+  expect_equivalent(set_contents(ht, 1, dplyr::matches("b"), 0), set_contents(ht, 1, 2, 0))
+
+  align(ht) <- "right"
+  test_props_same <- function(ht2) expect_equivalent(align(ht2), align(ht))
+  test_props_same(set_contents(ht, 1:6))
+  test_props_same(set_contents(ht, 1:6, byrow = TRUE))
+  test_props_same(set_contents(ht, 1, 1, 0))
+  test_props_same(set_contents(ht, 1, 1, 0))
+  test_props_same(set_contents(ht, 2:3, 2, 3:2))
+})
+
 test_that("merge_cells", {
   ht <- hux(a = 1:3, b = 1:3)
   expect_silent(ht2 <- merge_cells(ht, 1, 1:2))
@@ -265,24 +235,3 @@ test_that("merge_cells", {
 
 })
 
-
-test_that("where", {
-  dfr <- data.frame(a = 1:3, b = letters[1:3], d = 3:1, stringsAsFactors = FALSE)
-  expect_equivalent(where(dfr == 3), matrix(c(3, 1, 1, 3), 2, 2))
-})
-
-
-test_that("where() works with set_*", {
-  ht <- hux(a = c(1, 0), b = c(0, 1))
-  ht <- set_font(ht, where(ht > 0), "times")
-  expect_equivalent(font(ht), matrix(c("times", NA, NA, "times"), 2, 2))
-})
-
-
-test_that("is_a_number", {
-  expect_false(is_a_number("foo"))
-  expect_true(is_a_number(1.5))
-  expect_true(is_a_number("1.5"))
-  ht <- hux(a = 1:2, add_colnames = TRUE)
-  expect_equivalent(is_a_number(ht), matrix(c(FALSE, TRUE, TRUE), 3, 1))
-})